Add strategy framework and configurable momentum strategy
